backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2                  # HTTP client used by ingest.py and models/ollama.py

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.10        # PostgreSQL driver
pgvector==0.3.6                # pgvector type registration for psycopg2

# ── PDF parsing ────────────────────────────────────────────────────────────────
pymupdf==1.25.1                # native text extraction (fitz)
pytesseract==0.3.13            # OCR for scanned pages
Pillow==11.1.0                 # image processing for pytesseract

# ── Retrieval ──────────────────────────────────────────────────────────────────
rank-bm25==0.2.2               # BM25 keyword search
langchain-text-splitters==0.3.8  # RecursiveCharacterTextSplitter used in chunker.py
